+++ lib/karafka/routing/consumer_group.rb
@@ -31,10 +31,7 @@ module Karafka
 
       # @return [Boolean] true if this consumer group should be active in our current process
       def active?
-        cgs = Karafka::App.config.internal.routing.active.consumer_groups
-
-        # When empty it means no groups were specified, hence all should be used
-        cgs.empty? || cgs.include?(name)
+        Karafka::App.config.internal.routing.activity_manager.active?(:consumer_groups, name)
       end
 
       # Builds a topic representation inside of a current consumer group route