+++ spec/integrations/routing/routing_and_cli_validations_for_exclusions_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
-# Karafka should fail when defined routing is invalid
-# Karafka should fail if we want to listen on a topic that was not defined
+# Karafka should fail when excluding non existing routing elements
 
 setup_karafka
 
@@ -34,7 +33,7 @@ rescue Karafka::Errors::InvalidConfigurationError
 end
 
 ARGV[0] = 'server'
-ARGV[1] = '--consumer-groups'
+ARGV[1] = '--exclude-consumer-groups'
 ARGV[2] = 'non-existing'
 
 Karafka::Cli.prepare
@@ -48,11 +47,12 @@ rescue Karafka::Errors::InvalidConfigurationError => e
 end
 
 ARGV.clear
+Karafka::App.config.internal.routing.activity_manager.clear
 
 assert_equal 3, guarded.size
 
 ARGV[0] = 'server'
-ARGV[1] = '--subscription-groups'
+ARGV[1] = '--exclude-subscription-groups'
 ARGV[2] = 'non-existing'
 
 Karafka::Cli.prepare
@@ -66,11 +66,12 @@ rescue Karafka::Errors::InvalidConfigurationError => e
 end
 
 ARGV.clear
+Karafka::App.config.internal.routing.activity_manager.clear
 
 assert_equal 4, guarded.size
 
 ARGV[0] = 'server'
-ARGV[1] = '--topics'
+ARGV[1] = '--exclude-topics'
 ARGV[2] = 'non-existing'
 
 Karafka::Cli.prepare
@@ -84,5 +85,6 @@ rescue Karafka::Errors::InvalidConfigurationError => e
 end
 
 ARGV.clear
+Karafka::App.config.internal.routing.activity_manager.clear
 
 assert_equal 5, guarded.size